made heat int from float

+++ worldBuilder/world.py
@@ -381,7 +381,7 @@ class World():
             self._temp -= 1
         
         if self._heatSource >= 1:
-            self._temp += math.pow(self._heatSource,1.5)
+            self._temp += int(math.pow(self._heatSource,1.5))
         
         ## Shrinking of heat source
         if self._prevHeatSource == self._heatSource:
